Fix gcc warnings due to previous commits
git-svn-id: https://svn.osgeo.org/gdal/trunk@31598 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogr_srs_panorama.cpp
@@ -141,7 +141,7 @@ static const long aoEllips[] =
     7003   // Australian National, 1965
 };
 
-#define NUMBER_OF_ELLIPSOIDS    (sizeof(aoEllips)/sizeof(aoEllips[0]))
+#define NUMBER_OF_ELLIPSOIDS    static_cast<int>(sizeof(aoEllips)/sizeof(aoEllips[0]))
 
 /************************************************************************/
 /*                        OSRImportFromPanorama()                       */
@@ -432,7 +432,7 @@ OGRErr OGRSpatialReference::importFromPanorama( long iProjSys, long iDatum,
         }
 
         else if ( iEllips > 0
-                  && iEllips < (long)NUMBER_OF_ELLIPSOIDS
+                  && iEllips < NUMBER_OF_ELLIPSOIDS
                   && aoEllips[iEllips] )
         {
             char    *pszName = NULL;
